+++ tests/sql_mock/clickhouse/test_column_mocks.py
@@ -1,9 +1,9 @@
-from sql_mock.clickhouse.column_mocks import Array, ClickhouseColumnMock, Decimal
+from sql_mock.clickhouse.column_mocks import Array, ClickhouseColumnMock, Decimal, String, Int
 
 
 def test_init_not_nullable():
     """
-    ...then nullable should be False and dtype be the same as passed.
+    ...then nullable should be True and dtype be the same as passed.
     """
 
     class ColMock(ClickhouseColumnMock):
@@ -18,7 +18,7 @@ def test_init_not_nullable():
 
 def test_init_nullable():
     """
-    ...then nullable should be True and dtype should be wrapped with "Nullable()"
+    ...then nullable should be False and dtype should be wrapped with "Nullable()"
     """
 
     class ColMock(ClickhouseColumnMock):
@@ -49,8 +49,8 @@ def test_decimal_initialization_nullable():
 
 def test_array_column_inner_dtype():
     """Ensure that the inner dtype is processed correctly"""
-    string_array_col = Array(inner_dtype="String", default=["a", "b"], nullable=True)
-    int_array_col = Array(inner_dtype="Integer", default=[1, 2], nullable=False)
+    string_array_col = Array(inner_type=String, default=["a", "b"], nullable=True)
+    int_array_col = Array(inner_type=Int, default=[1, 2], nullable=False)
 
     assert string_array_col.dtype == "Nullable(Array(String))"
     assert string_array_col.default == ["a", "b"]